+++ components/alert.js
@@ -27,7 +27,8 @@ export default function Alert({ preview }) {
             <>
               The source code for this blog is{' '}
               <a
-                href={`https://github.com/vercel/next.js/tree/canary/examples/${EXAMPLE_PATH}`}
+                href={`https://github.com/terwer/cms-wordpress-app`}
+                target="_blank"
                 className="underline hover:text-success duration-200 transition-colors"
               >
                 available on GitHub
