+++ lib/constants.js
@@ -1,4 +1,3 @@
-export const EXAMPLE_PATH = 'cms-wordpress'
 export const CMS_NAME = 'WordPress'
 export const CMS_URL = 'https://wordpress.org'
 export const HOME_OG_IMAGE_URL =
